+++ Changelog.md
@@ -5,6 +5,7 @@
 * Add a new theme, "Dawn".
 * Add Distance Conversions.
 * Add Volume Conversions.
+* In the Quadratic Equation Solver, pressing Enter in one of the input fields has the same effect as clicking the Solve button.
 * Update the instructions in the Quadratic Equation Solver.
 * Update the interface of the Quadratic Equation Solver.
 * Update the interface of the Triangles window.